fix: show sidebar on sandbox-policies page without moving to settings
Keep page at /sandbox-policies URL. Render SettingsSidebar on desktop
only (hidden md:block). Update sidebar nav href to /sandbox-policies.
/settings/sandbox-policies now redirects back to /sandbox-policies.
